feat: reorganize affix+priorities sections
